+++ packages/effector-vue/options-vue3.d.ts
@@ -1,3 +1,16 @@
-import { Plugin } from 'vue'
+import {Plugin} from 'vue'
+import {Effect, Event, Store} from 'effector'
 
-declare function VueEffector(plugin: Plugin): void
+type EffectorOptionValue =
+  | Store<any>
+  | Event<any>
+  | Effect<any, any, any>
+  | {value: Store<any>; vModel?: boolean}
+
+declare module 'vue' {
+  interface ComponentCustomOptions {
+    effector?: () => Record<string, EffectorOptionValue>
+  }
+}
+
+export declare const VueEffector: Plugin